Desarrollo

+++ Backend/app/modules/ahorro/application/use_cases/eliminar_ahorro.py
@@ -1,7 +1,24 @@
+from app.shared.exceptions.business_exceptions import (
+    AhorroNoEncontrado,
+    CuentaNoEncontrada,
+)
+
+
 class EliminarAhorroUseCase:
 
     def __init__(self, repository):
         self.repository = repository
 
-    def execute(self, id_ahorro):
+    def execute(self, id_ahorro, id_usuario):
+
+        cuenta = self.repository.get_cuenta_por_usuario(id_usuario)
+
+        if not cuenta:
+            raise CuentaNoEncontrada()
+
+        ahorro = self.repository.get_by_id(id_ahorro)
+
+        if not ahorro or ahorro.id_cuenta != cuenta.id_cuenta:
+            raise AhorroNoEncontrado()
+
         return self.repository.delete(id_ahorro)
